fix: address api head review feedback
Extract Allow header generation into a shared helper so production and dev dispatch stay in sync.

Cancel discarded response bodies when HEAD handling suppresses streaming responses and extend regression coverage for that cleanup.
